+++ payments/src/components/Button.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const Button = () => {
+return (
+  	<div className="button-box">
+      <button type="button" className="button-text">다음</button>
+    </div>
+);
+}
+export default Button;
