+++ payments/src/components/CardNumberInput.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const CardNumberInput = () => {
+return (
+    <>
+      <div className="input-box">
+        <input className="input-basic" type="text" />
+        <input className="input-basic" type="text" />
+        <input className="input-basic" type="password" />
+        <input className="input-basic" type="password" />
+      </div>
+    </>
+);
+}
+export default CardNumberInput;
