+++ payments/src/components/CardSecret.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const CardSecret = () => {
+return (
+   <>
+   	  <input className="input-basic w-15" type="password" />
+					<input className="input-basic w-15" type="password" />
+					<input className="input-basic w-15" type="password" />
+					<input className="input-basic w-15" type="password" />
+   </>
+);
+}
+export default CardSecret;
