+++ payments/src/components/DateInput.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const DateInput = () => {
+return (
+   	<div className="input-box w-50">
+						<input className="input-basic" type="text" placeholder="MM" />
+						<input className="input-basic" type="text" placeholder="YY" />
+					</div>
+);
+}
+export default DateInput;
