+++ payments/src/components/Input.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const Input = () => {
+return (
+	// text type
+   	<input
+						type="text"
+						className="input-basic"
+						placeholder="카드에 표시된 이름과 동일하게 입력하세요."
+					/>
+					// password type
+					//	<input className="input-basic w-25" type="password" />
+)}
+export default Input;
